Support filtering based on series number / suffix
We add support for two common scenarios:

* series which result in multiple output niftis from dcm2niix (e.g., multi-echo
 sequences where individual echoes are written out in addition to the final
 image, even though these component files are generally of no interest)

* cases where particular runs should be ignored for a given subject / session
 (e.g., 'bad' task fMRI runs, structural scans that were re-acquired due to head
 motion)

We achieve this by adding a new input argument (selectseries), and parsing of
two new sub-fields in the 'criteria' subfield of the configuration json:
seriesnum and suffix.

The idea with the input argument is that ignored series generally are
ideosyncratic to a particular subject/session, so it makes sense to set this at
call time to avoid having to generate custom config files for each scan session.

The idea with the additional json criteria fields is to suppress scans that are
never of interest. For instance, some centers run a 'test' EPI acquisition
routinely before data collection begins, so seriesnum 1 is always junk.

Tested on Python 2.7 and 3.6.