Correct handling of tasks, custom sidecar fields
BIDS specifies that tasks should be encoded as a TaskName field in the sidecar
(and also in the field name with the task- prefix). This commit adds support for
sidecar customization with the customHeader field in the configuration file. If
TaskName is set, we rename the nifti appropriately.

customHeader also makes it possible to manually add in other information to the
fairly sparse sidecar provided by dcm2niix (e.g., acceleration information for
parallel imaging sequences).

Minor changes:
* update README with up-to-date usage info
* Acquisition.customLabels now inits to "" instead of None, which makes for neater conditionals throughout
* updated example/config.json to illustrate TaskName usage