Edit nifti header to reflect custom RepetitionTime
RepetitionTime is not always set correctly by dcm2niix. When you correct this
value in the sidecar (using the customHeader field in the description) you also
need to update the nifti header to avoid introducing conflicting information.

So we import nibabel (in a try block so it's only an optional dependency), and
make the change.

I also made a minor tweak to the logging behavior so that log messages from
different modules go to the same log file.