updating anonymiser process - new version of pydeface
Fix for new version of pydeface, where "--outfile" flag is needed if one needs to indicate the outfile. Related to #38.

Potentially the existing solution is more flexible and users would simply need to wrap their anonymiser in a shell script that has the required outputs/inputs. In this case it would be good to improve documentation.