Drop run_shell_command in favour of subprocess.run
run_shell_command was parsing the command line passed, so it was subject to shell injections
and, on Windows, wouldn't handle filenames correctly -- it would split on spaces (which are common
in Windows filenames) and interpret \s (also common in Windows file names) as control codes.

To do this, I got rid of defaceTpl, which was a shell command template, in favour of hardcoding
pydeface as the anonymizing algorithm, and as a result of that I also bumped the major version.